add validation for input data, add global constants messages

+++ Dev_Ocean/Services/ISpacehipsService.cs
@@ -4,6 +4,6 @@ namespace DevOcean.Services
 {
     public interface ISpaceshipsService
     {
-        public double CalculateTax(Spaceship spaceship, int yearOfCalculation);
+        public int CalculateTax(Spaceship spaceship, int yearOfCalculation);
     }
 }
