+++ Dev_Ocean/Services/SpaceshipsService.cs
@@ -4,7 +4,7 @@ namespace DevOcean.Services
 {
     public class SpaceshipsService : ISpaceshipsService
     {
-        public double CalculateTax(Spaceship spaceship, int yearOfCalculation)
+        public int CalculateTax(Spaceship spaceship, int yearOfCalculation)
         {
             return spaceship.Tax + (spaceship.MilesTraveled / 1000) * spaceship.TaxIncreaser - (yearOfCalculation - spaceship.YearOfPurchase) * spaceship.TaxDecreaser;
